メソッド名の変更完了

+++ app/javascript/views/admin/FixedShiftsIndex.vue
@@ -141,8 +141,10 @@
       },
 
       deleteFixedShiftInTableData(selectedShiftID, index) {
+        if (window.confirm(`このシフトを削除します、よろしいでしょうか?`)) {
         this.$store.state.fixedShiftsInTableData.splice(index, 1)
         this.$store.dispatch( 'deleteFixedShiftInTableData', {shiftID: selectedShiftID, type: "fixed"} )
+        }
       },
 
       updateFixedShiftsInTableData() {
